+++ src/firecracker/src/main.rs
@@ -22,7 +22,7 @@ use vmm::signal_handler::register_signal_handlers;
 use vmm::version_map::{FC_VERSION_TO_SNAP_VERSION, VERSION_MAP};
 use vmm::vmm_config::instance_info::{InstanceInfo, VmState};
 use vmm::vmm_config::logger::{init_logger, LoggerConfig, LoggerLevel};
-use vmm::{resources::VmResources, EventManager, ExitCode};
+use vmm::{resources::VmResources, EventManager, ExitCode, HTTP_MAX_PAYLOAD_SIZE};
 
 // The reason we place default API socket under /run is that API socket is a
 // runtime file.
@@ -102,6 +102,8 @@ fn main_exitable() -> ExitCode {
             error!("Failed to write metrics while panicking: {}", e);
         }
     }));
+
+    let http_max_payload_size_str = HTTP_MAX_PAYLOAD_SIZE.to_string();
 
     let mut arg_parser = ArgParser::new()
         .arg(
@@ -204,7 +206,12 @@ fn main_exitable() -> ExitCode {
         .arg(
             Argument::new("http-api-max-payload-size")
                 .takes_value(true)
-                .help("Http API request payload max size.")
+                .default_value(&http_max_payload_size_str)
+                .help("Http API request payload max size, in bytes.")
+        ).arg(
+            Argument::new("mmds-size-limit")
+                .takes_value(true)
+                .help("Mmds data store limit, in bytes.")
         );
 
     let arguments = match arg_parser.parse_from_cmdline() {
@@ -307,13 +314,26 @@ fn main_exitable() -> ExitCode {
 
     let boot_timer_enabled = arguments.flag_present("boot-timer");
     let api_enabled = !arguments.flag_present("no-api");
-    let payload_limit = arg_parser
+    let api_payload_limit = arg_parser
         .arguments()
         .single_value("http-api-max-payload-size")
         .map(|lim| {
             lim.parse::<usize>()
                 .expect("'http-api-max-payload-size' parameter expected to be of 'usize' type.")
-        });
+        })
+        // Safe to unwrap as we provide a default value.
+        .unwrap();
+
+    // If the mmds size limit is not explicitly configured, default to using the
+    // `http-api-max-payload-size` value.
+    let mmds_size_limit = arg_parser
+        .arguments()
+        .single_value("mmds-size-limit")
+        .map(|lim| {
+            lim.parse::<usize>()
+                .expect("'mmds-size-limit' parameter expected to be of 'usize' type.")
+        })
+        .unwrap_or_else(|| api_payload_limit);
 
     if api_enabled {
         let bind_path = arguments
@@ -345,7 +365,8 @@ fn main_exitable() -> ExitCode {
             instance_info,
             process_time_reporter,
             boot_timer_enabled,
-            payload_limit,
+            api_payload_limit,
+            mmds_size_limit,
             metadata_json.as_deref(),
         )
     } else {
@@ -358,7 +379,7 @@ fn main_exitable() -> ExitCode {
             vmm_config_json,
             instance_info,
             boot_timer_enabled,
-            payload_limit,
+            mmds_size_limit,
             metadata_json.as_deref(),
         )
     }
@@ -440,11 +461,11 @@ fn build_microvm_from_json(
     config_json: String,
     instance_info: InstanceInfo,
     boot_timer_enabled: bool,
-    mmds_max_size: Option<usize>,
+    mmds_size_limit: usize,
     metadata_json: Option<&str>,
 ) -> std::result::Result<(VmResources, Arc<Mutex<vmm::Vmm>>), ExitCode> {
     let mut vm_resources =
-        VmResources::from_json(&config_json, &instance_info, mmds_max_size, metadata_json)
+        VmResources::from_json(&config_json, &instance_info, mmds_size_limit, metadata_json)
             .map_err(|err| {
                 error!("Configuration for VMM from one single json failed: {}", err);
                 vmm::FC_EXIT_CODE_BAD_CONFIGURATION
@@ -473,7 +494,7 @@ fn run_without_api(
     config_json: Option<String>,
     instance_info: InstanceInfo,
     bool_timer_enabled: bool,
-    mmds_max_size: Option<usize>,
+    mmds_size_limit: usize,
     metadata_json: Option<&str>,
 ) -> ExitCode {
     let mut event_manager = EventManager::new().expect("Unable to create EventManager");
@@ -490,7 +511,7 @@ fn run_without_api(
         config_json.unwrap(),
         instance_info,
         bool_timer_enabled,
-        mmds_max_size,
+        mmds_size_limit,
         metadata_json,
     ) {
         Ok((res, vmm)) => (res, vmm),