+++ src/mmds/src/data_store.rs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::token::{Error as TokenError, TokenAuthority};
-use crate::MAX_DATA_STORE_SIZE;
 use serde::{Deserialize, Serialize};
 use serde_json::{to_vec, Value};
 use std::fmt;
@@ -91,18 +90,23 @@ impl fmt::Display for Error {
     }
 }
 
+// Used for ease of use in tests.
 impl Default for Mmds {
     fn default() -> Self {
-        Mmds {
-            data_store: Value::default(),
-            token_authority: None,
-            is_initialized: false,
-            data_store_limit: MAX_DATA_STORE_SIZE,
-        }
+        Self::default_with_limit(51200)
     }
 }
 
 impl Mmds {
+    pub fn default_with_limit(data_store_limit: usize) -> Self {
+        Mmds {
+            data_store: Value::default(),
+            token_authority: None,
+            is_initialized: false,
+            data_store_limit,
+        }
+    }
+
     /// This method is needed to check if data store is initialized.
     /// When a PATCH request is made on an uninitialized Mmds structure this method
     /// should return a NotFound error.
@@ -168,12 +172,17 @@ impl Mmds {
         self.data_store_limit = data_store_limit;
     }
 
-    // We do not check data_store size here because a request with a body
-    // bigger than the imposed limit will be stopped by micro_http before
-    // reaching here.
-    pub fn put_data(&mut self, data: Value) {
-        self.data_store = data;
-        self.is_initialized = true;
+    pub fn put_data(&mut self, data: Value) -> Result<(), Error> {
+        // It is safe to unwrap because any map keys are all strings and
+        // we are using default serializer which does not return error.
+        if to_vec(&data).unwrap().len() > self.data_store_limit {
+            Err(Error::DataStoreLimitExceeded)
+        } else {
+            self.data_store = data;
+            self.is_initialized = true;
+
+            Ok(())
+        }
     }
 
     pub fn patch_data(&mut self, patch_data: Value) -> Result<(), Error> {
@@ -334,7 +343,8 @@ mod tests {
 
         let mut mmds_json = "{\"meta-data\":{\"iam\":\"dummy\"},\"user-data\":\"1522850095\"}";
 
-        mmds.put_data(serde_json::from_str(mmds_json).unwrap());
+        mmds.put_data(serde_json::from_str(mmds_json).unwrap())
+            .unwrap();
         assert!(mmds.check_data_store_initialized().is_ok());
 
         assert_eq!(mmds.get_data_str(), mmds_json);
@@ -365,7 +375,7 @@ mod tests {
             "balance": -24
         }"#;
         let data_store: Value = serde_json::from_str(data).unwrap();
-        mmds.put_data(data_store);
+        mmds.put_data(data_store).unwrap();
 
         // Test invalid path.
         assert_eq!(
@@ -507,7 +517,6 @@ mod tests {
     #[test]
     fn test_update_data_store() {
         let mut mmds = Mmds::default();
-        mmds.set_data_store_limit(MAX_DATA_STORE_SIZE);
 
         let data = r#"{
             "name": {
@@ -517,7 +526,7 @@ mod tests {
             "age": "43"
         }"#;
         let data_store: Value = serde_json::from_str(data).unwrap();
-        mmds.put_data(data_store);
+        mmds.put_data(data_store).unwrap();
 
         let data = r#"{
             "name": {
@@ -537,7 +546,7 @@ mod tests {
             "age": 43
         }"#;
         let data_store: Value = serde_json::from_str(data).unwrap();
-        mmds.put_data(data_store);
+        mmds.put_data(data_store).unwrap();
 
         let data = r#"{
             "name": {
@@ -574,6 +583,22 @@ mod tests {
         assert!(mmds.patch_data(data_store).is_ok());
         assert!(mmds.get_data_str().contains("smth2"));
         assert_eq!(mmds.get_data_str().len(), 72);
+    }
+
+    #[test]
+    fn test_put_size_limit() {
+        let mut mmds = Mmds::default();
+        let filling = (0..51300).map(|_| "X").collect::<String>();
+        let data = "{\"key\": \"".to_string() + &filling + "\"}";
+
+        let data_store: Value = serde_json::from_str(&data).unwrap();
+
+        assert_eq!(
+            mmds.put_data(data_store).unwrap_err().to_string(),
+            Error::DataStoreLimitExceeded.to_string()
+        );
+
+        assert_eq!(mmds.get_data_str().len(), 2);
     }
 
     #[test]